heretic 2 callbacks

+++ src/common/header/shared.h
@@ -644,7 +644,7 @@ typedef struct
 {
 	pmtype_t pm_type;
 
-	short origin[3];            /* 12.3 */
+	int origin[3];            /* 12.3 */
 	short velocity[3];          /* 12.3 */
 	byte pm_flags;              /* ducked, jump_held, etc */
 	byte pm_time;               /* each unit = 8 ms */
